backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
psycopg2-binary

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]

# Validation & Settings
pydantic
pydantic-settings
email-validator

# Development
python-dotenv

# AI/ML Libraries
openai>=1.0.0
anthropic
langchain
langchain-openai
transformers
torch
sentence-transformers
numpy
pandas
openpyxl
xlrd
scikit-learn

# NLP & Sentiment Analysis
nltk
textblob
vaderSentiment

# News & Data Collection
feedparser
beautifulsoup4
requests

# AWS Services
boto3
botocore

# Caching
redis
hiredis
